+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Optional truncate native trace
 * Opts: Global options of the library
 * InvalidConfig::getErrmsg() is deprecated. Use getErrorMessage(). 
+* NotValid::getErrmsg() is deprecated. Use getErrorMessage().
 
 ## 0.0.5 (18.02.2014)